ocrchestra_django/requirements.txt
Django==5.0
djangorestframework==3.14.0
drf-spectacular==0.29.0
Celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0

# Existing OCRchestra dependencies
jsonschema
pytest
requests==2.31.0
PyMuPDF==1.23.8
python-docx==1.1.0
Pillow==10.1.0
sqlalchemy
groq
pdfplumber==0.10.3
pandas
plotly
